indexer/requirements.txt
azure-search-documents==11.6.0b5
azure-storage-blob
azure-identity